refactor: split content into data and routes
