+++ app/about/careers/page.tsx
@@ -6,12 +6,12 @@ import {
   ContentTitle,
 } from "@/components/ContentSection"
 import { ButtonLink } from "@/components/button/ButtonLink"
-import CareersContent from "@/content/about/careers.mdx"
+import CareersContent from "@/content/routes/about/careers.mdx"
 import { CareerData } from "@/components/CareerData"
 import { getMDXMetadata } from "@/lib/mdx-utils"
 
 export default async function Careers() {
-  const metadata = getMDXMetadata("content/about/careers.mdx")
+  const metadata = getMDXMetadata("content/routes/about/careers.mdx")
   return (
     <>
       <Hero
