+++ app/about/grant-and-publication-materials/page.tsx
@@ -1,11 +1,11 @@
 import React from "react"
 import { Hero } from "@/components/Hero"
-import GrantAndPublicationMaterialsContent from "@/content/about/grant-and-publication-materials.mdx"
+import GrantAndPublicationMaterialsContent from "@/content/routes/about/grant-and-publication-materials.mdx"
 import { getMDXMetadata } from "@/lib/mdx-utils"
 
 export default async function PublicationAndGrantMaterials() {
   const metadata = getMDXMetadata(
-    "content/about/grant-and-publication-materials.mdx"
+    "content/routes/about/grant-and-publication-materials.mdx"
   )
 
   return (
