+++ app/about/help/page.tsx
@@ -1,10 +1,10 @@
 import React from "react"
-import ContactContent from "@/content/about/help.mdx"
+import ContactContent from "@/content/routes/about/help.mdx"
 import { getMDXMetadata } from "@/lib/mdx-utils"
 import { Hero } from "@/components/Hero"
 
 export default async function ContactUs() {
-  const metadata = getMDXMetadata("content/about/help.mdx")
+  const metadata = getMDXMetadata("content/routes/about/help.mdx")
 
   return (
     <>
