+++ app/ai/ai-tools/page.tsx
@@ -1,10 +1,10 @@
 import { Hero } from "@/components/Hero"
 import { ButtonLink } from "@/components/button/ButtonLink"
-import AIToolsContent from "@/content/ai/ai-tools.mdx"
+import AIToolsContent from "@/content/routes/ai/ai-tools.mdx"
 import { getMDXMetadata } from "@/lib/mdx-utils"
 
 export default function AITools() {
-  const metadata = getMDXMetadata("content/ai/ai-tools.mdx")
+  const metadata = getMDXMetadata("content/routes/ai/ai-tools.mdx")
 
   return (
     <>
