+++ app/page.tsx
@@ -16,21 +16,23 @@ import {
   ContentTitle,
 } from "@/components/ContentSection"
 import { FaCalendarAlt } from "react-icons/fa"
+import { getMDXMetadata } from "@/lib/mdx-utils"
 
 export default async function Home() {
   // Load featured carousel data from YAML
   const featuredCarouselRaw = await readContentFile<{
     carousel: StyledCarouselItem[]
-  }>("content/home/featured-carousel.yaml")
+  }>("content/data/featured-carousel.yaml")
   const featuredCarouselData = featuredCarouselRaw.data.carousel
+  const metadata = getMDXMetadata("content/routes/home.mdx")
 
   try {
     return (
       <>
         <MainHero
           image={"/images/hero/ccv-original.webp"}
-          title="Center for Computation and Visualization"
-          description="Advancing computational research with scientific and computing expertise."
+          title={metadata.title}
+          description={metadata.description}
         >
           <ButtonLink
             variant="primary_filled"
