+++ app/services/department-support/page.tsx
@@ -1,9 +1,11 @@
 import { Hero } from "@/components/Hero"
-import DepartmentSupportContent from "@/content/services/department-support.mdx"
+import DepartmentSupportContent from "@/content/routes/services/department-support.mdx"
 import { getMDXMetadata } from "@/lib/mdx-utils"
 
 export default function DepartmentSupport() {
-  const metadata = getMDXMetadata("content/services/department-support.mdx")
+  const metadata = getMDXMetadata(
+    "content/routes/services/department-support.mdx"
+  )
 
   return (
     <>
