+++ app/services/oscar/page.tsx
@@ -1,10 +1,10 @@
 import { Hero } from "@/components/Hero"
-import OscarContent from "@/content/services/oscar.mdx"
+import OscarContent from "@/content/routes/oscar.mdx"
 import { getMDXMetadata } from "@/lib/mdx-utils"
 import { ButtonLink } from "@/components/button/ButtonLink"
 
 export default async function Oscar() {
-  const metadata = getMDXMetadata("content/services/oscar.mdx")
+  const metadata = getMDXMetadata("content/routes/services/oscar.mdx")
 
   return (
     <>
