+++ content/routes/about/us.mdx
@@ -2,6 +2,8 @@
 title: About Us
 image: "/images/hero/about-kayaks.webp"
 description: The Center for Computation and Visualization provides high-performance computing and visualization services to the Brown community and also collaborates with researchers on projects across a vast range of disciplines.
+dataRefs:
+    - "data/people.yaml"
 ---
 
 <ContentSection title="Office of Information Technology" id="oit">
@@ -30,7 +32,7 @@ description: The Center for Computation and Visualization provides high-performa
 </ContentSection>
 
 <ContentSection title="Our Team" id="our-team">
-  <PeopleSection peopleContentPath="content/about/people.yaml" />
+  <PeopleSection peopleContentPath="content/data/people.yaml" />
 </ContentSection>
 
 <LocationSection title="Our Home" />
