+++ content/routes/ai/ai-oscar.mdx
