+++ content/routes/home.mdx
@@ -1,0 +1,6 @@
+---
+title: Center for Computation and Visualization
+description: Advancing computational research with scientific and computing expertise.
+dataRefs:
+    - "data/featured-carousel.yaml"
+---
