+++ content/routes/services/classroom.mdx
@@ -2,6 +2,8 @@
 title: "Classroom Support"
 description: |
   We support faculty in the classroom, providing tutorials or granting access to cutting edge technology for teaching with code.
+dataRefs:
+    - "data/classroom-carousel.json"
 ---
 
 <ContentSection title="Student Accounts" id="student-accounts">
