+++ content/routes/services/storage.mdx
@@ -1,4 +1,7 @@
 ---
 title: Storage and Transfer
 description: We provide a variety of services to allow you to share and store data.
+dataRefs:
+    - "data/storage-features.json"
+    - "data/storage-questions.json"
 ---
