+++ content/routes/services/stronghold.mdx
